// src/app/ws/soccer/model/NewsDTO.ts
/**
 * Soccer API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * Contact: voetbalsvk@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import * as models from "./models";

export interface NewsDTO {
    comments?: Array<models.CommentDTO>;

    content?: string;

    editable?: boolean;

    header?: string;

    id?: number;

    postDate?: string;

    postedBy?: models.AccountDTO;

    type?: NewsDTO.TypeEnum;

}
export namespace NewsDTO {
    export enum TypeEnum {
        SEND = <any> 'SEND',
        POST = <any> 'POST',
        POSTANDSEND = <any> 'POST_AND_SEND'
    }
}